+++ src/commonMain/kotlin/CallbackKSLog.kt
@@ -7,3 +7,7 @@ class CallbackKSLog(
 ) : KSLog {
     override fun performLog(level: LogLevel, tag: String?, message: Any, throwable: Throwable?) = performLogCallback(level, tag, message, throwable)
 }
+
+fun KSLog(
+    performLogCallback: SimpleKSLogCallback
+) = CallbackKSLog(performLogCallback)