+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add `setDefaultKSLog` and `addDefaultKSLog`
 * Add `plus` operation for two `KSLog` instances to call them both on each log performing
 * Add `KSLog` factory for simple creation of `CallbackKSLog`
+* Add `MessageFormatter` factory
 
 ## 0.4.1